requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
griffe==1.6.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
numpy==1.26.4
opencv-python==4.10.0.84
openai==1.68.2
openai-agents==0.0.6
pillow==10.3.0
playwright==1.51.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
